Move view model
