+++ src/Files.App/GlobalUsings.cs
@@ -51,6 +51,7 @@ global using global::Files.App.Services;
 global using global::Files.App.UserControls;
 global using global::Files.App.UserControls.TabBar;
 global using global::Files.App.ViewModels;
+global using global::Files.App.ViewModels.Shells;
 global using global::Files.App.ViewModels.UserControls;
 global using global::Files.App.Views;
 global using global::Files.App.Views.Layouts;
